First Version of Nav Mesh
Got the first version of the nav mesh working by creating a graph of
walkable tiles.. I added a debug drawer class to the project so I can
draw the mesh to the screen. Looks real pretty lol. Currently its very
inefficient, because it connects every node with every node in a 5 tile
radius and probably duplicates most of the edges. Also it takes 5 sec
longer to start game.