Game menu works but I cannot play the game.
Using the starterkit as the base for the menu, I noticed that the screen
manager calls a file called GameplayScreen.cs and I simply tried to
merge our PiN_Game.cs code into that file.  The problem is with Services
and Content in PiN_Game conflicting with ScreenManager.Game.Services and
content in GameplayScreen.cs. The menu works but choosing the 'Play
Game' option will crash the game. To run the original pin_game file,
simply change the new instance in Game.cs file.